wip: various things
Create Week & MonthlyPlanner struct.
Create utility functions for adding & substracting day(s) to a date.
Function to create week based on the first or last day.

I'm now able to create all the normal weeks based on a date range.
Next step is to test the range function and create a wrapper function to package everything instead of calling things in the main function.